+++ local/src/main/java/com/woowahan/accountbook/local/datasource/HistoryDataSourceImpl.kt
@@ -49,7 +49,7 @@ class HistoryDataSourceImpl @Inject constructor(
 
     override suspend fun insertHistory(
         date: Long,
-        amount: Int,
+        amount: Long,
         content: String,
         category: Category,
         paymentMethod: PaymentMethod?
@@ -60,7 +60,7 @@ class HistoryDataSourceImpl @Inject constructor(
     override suspend fun updateHistory(
         id: Int,
         date: Long,
-        amount: Int,
+        amount: Long,
         content: String,
         category: Category,
         paymentMethod: PaymentMethod?
